Move default_item_attributes functions to external.database.datamodels.
