+++ components/server/src/external/routes/metric.py
@@ -5,7 +5,7 @@ from typing import Any
 import bottle
 from pymongo.database import Database
 
-from database.datamodels import default_metric_attributes, latest_datamodel
+from database.datamodels import latest_datamodel
 from database.measurements import insert_new_measurement, latest_measurement
 from database.reports import insert_new_report, latest_reports
 from model.actions import copy_metric, move_item
@@ -14,6 +14,7 @@ from model.metric import Metric
 from server_utilities.functions import uuid
 from server_utilities.type import MetricId, SubjectId
 
+from ..database.datamodels import default_metric_attributes
 from ..utils.functions import sanitize_html
 
 from .plugins.auth_plugin import EDIT_REPORT_PERMISSION
