+++ components/server/src/external/routes/subject.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta, timezone
 import bottle
 from pymongo.database import Database
 
-from database.datamodels import default_subject_attributes, latest_datamodel
+from database.datamodels import latest_datamodel
 from database.measurements import measurements_by_metric
 from database.reports import insert_new_report, latest_reports, metrics_of_subject
 from model.actions import copy_subject, move_item
@@ -12,6 +12,7 @@ from model.data import ReportData, SubjectData
 from server_utilities.functions import uuid
 from server_utilities.type import MetricId, ReportId, SubjectId
 
+from ..database.datamodels import default_subject_attributes
 from ..utils.functions import report_date_time, sanitize_html
 
 from .plugins.auth_plugin import EDIT_REPORT_PERMISSION
